// src/app/models/entity/location.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Address} from './address';
import {Room} from './room';


export interface Location {
  address?: Address;
  createdDate?: any;
  id?: number;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  name?: string;
  rooms?: Room[];
}